Refactor GUI code
