+++ src/main/java/duke/application/Launcher.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.application;
 
 import javafx.application.Application;
