+++ src/main/java/duke/application/MainWindow.java
@@ -1,9 +1,8 @@
-package duke;
+package duke.application;
 
-import duke.ui.Ui;
+import duke.Duke;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
-import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
